+++ src/app/Http/Controllers/UserController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Crypt;
 
@@ -11,16 +13,17 @@ class UserController extends Controller
     /**
      * Update the API key on the user's account.
      *
+     * @param Request $request
      * @return mixed
      */
-    public function updateApiKey()
+    public function updateApiKey(Request $request): RedirectResponse
     {
-        request()->validate([
+        $request->validate([
             'api_key' => 'required|string',
         ]);
 
         $user = User::find(Auth::user()->id);
-        $user->api_key = Crypt::encryptString(request()->get('api_key'));
+        $user->api_key = Crypt::encryptString($request->input('api_key'));
         $user->save();
 
         session()->flash('message', 'API Key updated');