+++ src/resources/views/modals/create-parcel.blade.php
@@ -15,16 +15,16 @@
         <hr>
 
         <label for="length">Length</label>
-        <input class="form-control" name="length" value="{{ old('length') }}">
+        <input class="form-control" name="length" type="number" step="0.1" value="{{ old('length') }}">
 
         <label for="width">Width</label>
-        <input class="form-control" name="width" value="{{ old('width') }}">
+        <input class="form-control" name="width" type="number" step="0.1" value="{{ old('width') }}">
 
         <label for="height">Height</label>
-        <input class="form-control" name="height" value="{{ old('height') }}">
+        <input class="form-control" name="height" type="number" step="0.1" value="{{ old('height') }}">
 
         <label for="weight">Weight</label>
-        <input class="form-control" name="weight" value="{{ old('weight') }}">
+        <input class="form-control" name="weight" type="number" step="0.1" value="{{ old('weight') }}">
 
         {{-- We do not close the form here because we do so at the component level --}}
 </x-modal>